poc/lyg/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intent-vector-hybrid-lab"
version = "0.1.0"
description = "Benchmark harness for intent-based embedding/vector/hybrid/rerank/filter pipelines"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
  "openai>=1.0.0",
  "google-genai>=1.0.0",
  "cohere>=5.0.0",
]

[project.scripts]
ivhl = "ivhl.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
